+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Parser
   - **Breaking Change**: `ParserBuilder.checkSymbols` is now private, use `options` in `ParserBuilder.build` instead.
   - **Breaking Change**: Remove `AdvancedBuilder.expand`, make `AdvancedBuilder` extend `ParserBuilder`, they both implement `IParserBuilder`.
+  - Add `ParserBuilder.use` to apply custom decorators with chain call.
 
 ## v0.7.0